remove old findowner
